Refactored tests into groups.
