+++ package.json
@@ -172,7 +172,7 @@
         "build": "node ./out/src/icon-manifest/build.js",
         "example": "node ./out/src/example/example.js ",
         "pretest": "npm run compile",
-        "test": "mocha ./out/test",
+        "test": "mocha ./out/test --recursive",
         "lint": "tslint src/**/*.ts test/**/*.ts",
         "test:vs": "node ./node_modules/vscode/bin/test",
         "vscode:prepublish": "rimraf ./out && npm run build",
